+++ CHANGELOG.md
@@ -12,6 +12,7 @@ a workaround to activate the extension when the module has no blocks yet
 - Added actions to compile/run GRC Flowgraph files or edit them in GNURadio Companion from the editor titlebar
 - Added a filter (regex) when creating Python bindings, making YAML from implementation, disabling or removing blocks
 - Added automatic blockname for modtool commands when a corresponding editor is in focus
+- Added an option to pick Python/C++ QA when creating blocks
 
 ### Changed
 - Replaced `child_process` with `python-shell` for modtool commands